requirements.txt
psycopg2
Django==3.2.5
djangorestframework==3.12.4
gunicorn==20.1.0
